% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoobyGrade.R
\name{ScoobyGrade}
\alias{ScoobyGrade}
\title{ScoobyGrade
This function can be used to visualize histograms plots inspired by the spooky cast of Scoopy Doo. Designed with grade distribution for assignments in mind.}
\usage{
ScoobyGrade(data, assignment, scoobCharac)
}
\arguments{
\item{data}{Inserts the data frame you'd like to use}

\item{assignment}{Inserts the column you would like to plot}

\item{scoobCharac}{Assign the color palette of a Scooby Doo crewmember for the histogram. Options are: Shaggy, Fred, Velma, Daphne, and Scooby.}
}
\description{
ScoobyGrade
This function can be used to visualize histograms plots inspired by the spooky cast of Scoopy Doo. Designed with grade distribution for assignments in mind.
}
\examples{
ScoobyGrade(classScores, "Final.exam", "Scooby")
}
