% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semPro_function_draft.R
\name{semPro}
\alias{semPro}
\title{semPro
This function can be used to visualize student progress academically over the course of one semester.}
\usage{
semPro(data, student)
}
\arguments{
\item{data}{Inserts the data frame you'd like to use}

\item{student}{Insert student name to receive a visual progress of student academic performance.}
}
\description{
semPro
This function can be used to visualize student progress academically over the course of one semester.
}
\examples{
semPro(grade, "Abreu, Jose")
}
