% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_grade.R
\name{weighted_grade}
\alias{weighted_grade}
\title{weighted_grade
This function is used to calculate the weighted average for students over the course of a semester.}
\usage{
weighted_grade(data, exam1, exam2, final, lab, students, prbSet)
}
\arguments{
\item{data}{Inserts the data frame you'd like to use}

\item{exam1}{Inserts the column (exam grades) you would like to take into account for overall grade}

\item{exam2}{Inserts the column (exam grades) you would like to take into account for overall grade}

\item{final}{Inserts the column (exam grades) you would like to take into account for overall grade}

\item{lab}{Inserts the column (lab score) you would like to take into account for overall grade}

\item{students}{Space to specify which row (students) you would like to calculate the average for to add into weighted grade. To call these, add in row range.}

\item{prbSet}{Space to specify which columns (problem sets) you would to like calculate the average for to add into weighted grade. To call these, add in column range...}
}
\description{
weighted_grade
This function is used to calculate the weighted average for students over the course of a semester.
}
\examples{
weighted_grade(grade, "Exam.1", "Exam.2", "Final.exam", "Lab", students=1:166, prbSet=7:18)
}
